requirements.txt
pandas==2.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.28
